Remove dead code, sort imports
