+++ apps/web/app/(authenticated)/(app)/app/apis/page.tsx
@@ -13,7 +13,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
 
 export default async function TenantOverviewPage() {
   const tenantId = getTenantId();
