+++ apps/web/app/(authenticated)/(app)/app/keys/CreateKey.tsx
@@ -1,13 +1,15 @@
 "use client";
 
-import { useState } from "react";
-import { useRouter } from "next/navigation";
 import { useToast } from "@/components/ui/use-toast";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
 
-import { trpc } from "@/lib/trpc/client";
 import { CopyButton } from "@/components/CopyButton";
+import { VisibleButton } from "@/components/VisibleButton";
 import { Loading } from "@/components/loading";
+import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
+import { Code } from "@/components/ui/code";
 import {
   Dialog,
   DialogContent,
@@ -17,10 +19,8 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
+import { trpc } from "@/lib/trpc/client";
 import { AlertTriangle } from "lucide-react";
-import { VisibleButton } from "@/components/VisibleButton";
-import { Code } from "@/components/ui/code";
 
 type Props = {
   apiId?: string;
