+++ apps/web/app/(authenticated)/(app)/app/keys/[keyId]/client.tsx
@@ -1,14 +1,10 @@
 "use client";
 import { PageHeader } from "@/components/PageHeader";
-import { Policy } from "@unkey/policies";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Label } from "@/components/ui/label";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { useToast } from "@/components/ui/use-toast";
-import { DeleteKeyButton } from "../DeleteKey";
 import { Badge } from "@/components/ui/badge";
-import { Trash } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import { useToast } from "@/components/ui/use-toast";
+import { Trash } from "lucide-react";
+import { DeleteKeyButton } from "../DeleteKey";
 
 const _allActions = ["create", "read", "update", "delete"];
