+++ apps/web/app/(authenticated)/(app)/layout.tsx
@@ -1,5 +1,5 @@
-import { ReactQueryProvider } from "./ReactQueryProvider";
 import { Toaster } from "@/components/ui/toaster";
+import { ReactQueryProvider } from "./ReactQueryProvider";
 
 export default function Layout({
   children,
