+++ apps/web/app/(authenticated)/(app)/onboarding/page.tsx
@@ -1,7 +1,8 @@
 import { getTenantId } from "@/lib/auth";
-import { db, schema, eq } from "@unkey/db";
+import { db, eq, schema } from "@unkey/db";
 import { redirect } from "next/navigation";
 import { Onboarding } from "./client";
+
 export default async function OnboardingPage() {
   const tenantId = getTenantId();
   const workspace = await db.query.workspaces.findFirst({
