+++ apps/web/app/(authenticated)/auth/sign-in/email-signin.tsx
@@ -1,12 +1,13 @@
 "use client";
 
-import * as React from "react";
 import { useSignIn } from "@clerk/nextjs";
+import * as React from "react";
 
+import { Loading } from "@/components/loading";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Loading } from "@/components/loading";
 import { useToast } from "@/components/ui/use-toast";
+
 export function EmailSignIn(props: { verification: (value: boolean) => void }) {
   const { signIn, isLoaded: signInLoaded } = useSignIn();
   const { toast } = useToast();
