+++ apps/web/app/(authenticated)/auth/sign-up/email-code.tsx
@@ -1,13 +1,14 @@
 "use client";
 
-import * as React from "react";
-import { useRouter } from "next/navigation";
 import { useSignUp } from "@clerk/nextjs";
+import { useRouter } from "next/navigation";
+import * as React from "react";
 
+import { Loading } from "@/components/loading";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Loading } from "@/components/loading";
 import { useToast } from "@/components/ui/use-toast";
+
 export function EmailCode() {
   const router = useRouter();
   const { toast } = useToast();
